refactor(core): extract shared atomic mark navigation
